refactor: removed unused files/functionality
